+++ main.js
@@ -1,13 +1,20 @@
-import Random from './dist/random.min.mjs';
+import Random from './src/random.js';
 
 const seed = 'helloWorld';
 const random = new Random(seed);
+
+const obj = {
+    prop1: 'hello',
+    prop2: 'world',
+    prop3: '!'
+}
 
 const table = [
     {
         random: random.gen(),
         number: random.number(),
         between: random.between(1, 100),
+        key: random.keyFromObject(obj),
     }
 ];