+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Random CHANGELOG
 
+### Version 1.2.0
+
+* new method ```keyFromObject```
+    * returns a random key from a passed in object
+    * added test
+
 ### Version 1.1.0
 
 * created and built tests using jest